CI: Add msvc (#3)
* CI: Add msvc

* meson.build: Skip GCC/Clang-only flags on MSVC

get_supported_arguments() doesn't reliably filter flags on MSVC (e.g.
-fstack-protector-strong passes the probe but emits D9002 at compile
time). Guard the extra_flags block with a compiler id check.


---------

+++ tests/00_test.c
@@ -2,7 +2,7 @@
 
 void set_hand(POKEVAL_Hand_5 *hand, int faces[POKEVAL_HAND_SIZE], int suits[POKEVAL_HAND_SIZE]) {
   for (int i = 0; i < POKEVAL_HAND_SIZE; ++i) {
-    hand->card[i].face_val = faces[i];
-    hand->card[i].suit = suits[i];
+    hand->card[i].face_val = (int8_t)faces[i];
+    hand->card[i].suit = (int8_t)suits[i];
   }
 }
